requirements.txt
Flask==3.1.0
mysql-connector-python==8.3.0  # Official package (latest version)
